requirements.txt
opencv-python==4.8.0.76
torch==2.0.1
timm==0.9.7
pandas==2.1.0
Pillow==10.0.0
transformers==4.34.0
sentencepiece==0.1.99
protobuf==4.24.3
openai-whisper==20230918
moviepy==1.0.3
pydub==0.25.1